+++ support-frontend/assets/helpers/abTests/lpPreviousGiving.js
@@ -11,32 +11,6 @@ const SUPPORT_AGAIN_HEADER_VARIANT_NAME = 'control';
 
 // ----- Helpers ----- //
 
-export function getLongMonth(date: Date) {
-  return date.toLocaleDateString('default', { month: 'long' });
-}
-
-function nth(d) {
-  if (d >= 11 && d <= 13) {
-    return 'th';
-  }
-  switch (d % 10) {
-    case 1:
-      return 'st';
-    case 2:
-      return 'nd';
-    case 3:
-      return 'rd';
-    default:
-      return 'th';
-  }
-}
-
-export function getDateWithOrdinal(date: Date) {
-  const dayOfMonth = date.getDate();
-
-  return `${dayOfMonth}${nth(dayOfMonth)}`;
-}
-
 export function isInSupportAgainHeaderVariant(acquisitionData: ReferrerAcquisitionData) {
   const { abTests } = acquisitionData;
